pyproject.toml
[tool.poetry]
name = "philapy"
version = "0.1.0"
description = ""
authors = ["redat00"]
packages = [
    { include = "philapy" }
]

[tool.poetry.scripts]
philapy = "philapy.cli:main"


[tool.poetry.dependencies]
python = ">=3.10,<3.11"
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.10.1"
selenium = "^4.4.3"
webdriver-manager = "^3.8.3"
python-telegram-bot = "^13.14"
redis = "^4.3.4"
pytype = "^2022.9.8"
pytest = "^7.1.3"
pytest-mock = "^3.8.2"
pytest-cov = "^3.0.0"
beautifulsoup4 = "^4.11.1"
click = "^8.1.3"
lxml = "^4.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
length_sort = false
combine_star = false
use_parentheses = true
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "First Party"
import_heading_thirdparty = "Third Party"
import_heading_localfolder = "Local"
force_sort_within_sections = false
from_first = false
